Tests/LibWeb/Text/expected/css/scrollable-overflow-recalc-svg-foreignobject-sticky.txt

sticky position stable across partial SVG relayout: true
relayouts performed: 0
overflow was recalculated: true
